pyproject.toml
[build-system]
requires = ["flit_core >=3.5,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "hydra-submitit-chunked"
version = "0.1.1"
description = "Hydra plugin using submitit that allows chunking jobs"
readme = "README.md"
requires-python = ">= 3.8"
license = {file = "LICENSE"}
authors = [{name = "Dominik Fuchsgruber", email = "d.fuchsgruber@tum.de"}]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python"
]
dependencies = [
  "hydra-core",
]

[project.optional-dependencies]
test = ["pytest"]

[project.urls]
homepage = "https://github.com/WodkaRHR/hydra-submitit-chunked-launcher"

[tool.flit.module]
name = "hydra_plugins.hydra_submitit_chunked_launcher"

[tool.flit.sdist]
exclude = ["tests/", "example/"]

[tool.pytest.ini_options]
testpaths = ["tests"]